Rename Plotter and Wallet services
- they collide with types with same names, exported from same package
