+++ packages/api-react/src/hooks/useGetNFTWallets.ts
@@ -1,4 +1,4 @@
-import { Wallet, WalletType } from '@chia-network/api';
+import type { Wallet, WalletType } from '@chia-network/api';
 import { useMemo } from 'react';
 
 import { useGetWalletsQuery } from '../services';
