+++ packages/api-react/src/services/wallet.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-param-reassign -- This file use Immer */
-import { CAT, DID, Farmer, NFT, OfferTradeRecord, Pool, Wallet, WalletType, toBech32m } from '@chia-network/api';
+import { CAT, DID, Farmer, NFT, OfferTradeRecord, Pool, WalletService, WalletType, toBech32m } from '@chia-network/api';
 import type {
   CalculateRoyaltiesRequest,
   CalculateRoyaltiesResponse,
@@ -8,6 +8,7 @@ import type {
   PlotNFT,
   PlotNFTExternal,
   Transaction,
+  Wallet,
   WalletBalance,
   WalletConnections,
 } from '@chia-network/api';
@@ -58,7 +59,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     walletPing: build.query<boolean, {}>({
       query: () => ({
         command: 'ping',
-        service: Wallet,
+        service: WalletService,
       }),
       transformResponse: (response: any) => response?.success,
     }),
@@ -66,7 +67,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     getLoggedInFingerprint: build.query<string | undefined, {}>({
       query: () => ({
         command: 'getLoggedInFingerprint',
-        service: Wallet,
+        service: WalletService,
       }),
       transformResponse: (response: any) => response?.fingerprint,
       providesTags: ['LoggedInFingerprint'],
@@ -82,7 +83,7 @@ export const walletApi = apiWithTag.injectEndpoints({
         try {
           const { data, error } = await fetchWithBQ({
             command: 'getWallets',
-            service: Wallet,
+            service: WalletService,
           });
 
           if (error) {
@@ -162,7 +163,7 @@ export const walletApi = apiWithTag.injectEndpoints({
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onWalletCreated',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getWallets,
         },
       ]),
@@ -176,14 +177,14 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ transactionId }) => ({
         command: 'getTransaction',
-        service: Wallet,
+        service: WalletService,
         args: [transactionId],
       }),
       transformResponse: (response: any) => response?.transaction,
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onTransactionUpdate',
-          service: Wallet,
+          service: WalletService,
           onUpdate: (draft, data, { transactionId }) => {
             const {
               additionalData: { transaction },
@@ -205,7 +206,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ walletId }) => ({
         command: 'getPwStatus',
-        service: Wallet,
+        service: WalletService,
         args: [walletId],
       }),
       /*
@@ -228,7 +229,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ walletId, fee }) => ({
         command: 'pwAbsorbRewards',
-        service: Wallet,
+        service: WalletService,
         args: [walletId, fee],
       }),
       invalidatesTags: [
@@ -249,7 +250,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ walletId, poolUrl, relativeLockHeight, targetPuzzleHash, fee }) => ({
         command: 'pwJoinPool',
-        service: Wallet,
+        service: WalletService,
         args: [walletId, poolUrl, relativeLockHeight, targetPuzzleHash, fee],
       }),
       invalidatesTags: [
@@ -267,7 +268,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ walletId, fee }) => ({
         command: 'pwSelfPool',
-        service: Wallet,
+        service: WalletService,
         args: [walletId, fee],
       }),
       invalidatesTags: [
@@ -285,7 +286,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ walletType, options }) => ({
         command: 'createNewWallet',
-        service: Wallet,
+        service: WalletService,
         args: [walletType, options],
       }),
       invalidatesTags: [
@@ -302,7 +303,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ walletId }) => ({
         command: 'deleteUnconfirmedTransactions',
-        service: Wallet,
+        service: WalletService,
         args: [walletId],
       }),
       invalidatesTags: (_result, _error, { walletId }) => [
@@ -330,7 +331,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ walletId }) => ({
         command: 'getWalletBalance',
-        service: Wallet,
+        service: WalletService,
         args: [walletId],
       }),
       transformResponse: (response) => {
@@ -351,27 +352,27 @@ export const walletApi = apiWithTag.injectEndpoints({
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onCoinAdded',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getWalletBalance,
         },
         {
           command: 'onCoinRemoved',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getWalletBalance,
         },
         {
           command: 'onPendingTransaction',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getWalletBalance,
         },
         {
           command: 'onOfferAdded',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getWalletBalance,
         },
         {
           command: 'onOfferUpdated',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getWalletBalance,
         },
       ]),
@@ -380,17 +381,17 @@ export const walletApi = apiWithTag.injectEndpoints({
     getFarmedAmount: build.query<any, undefined>({
       query: () => ({
         command: 'getFarmedAmount',
-        service: Wallet,
+        service: WalletService,
       }),
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onCoinAdded',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getFarmedAmount,
         },
         {
           command: 'onCoinRemoved',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getFarmedAmount,
         },
       ]),
@@ -470,7 +471,7 @@ export const walletApi = apiWithTag.injectEndpoints({
                 subscribeResponse = await baseQuery(
                   {
                     command: 'onTransactionUpdate',
-                    service: Wallet,
+                    service: WalletService,
                     args: [
                       (data: any) => {
                         const {
@@ -490,7 +491,7 @@ export const walletApi = apiWithTag.injectEndpoints({
               // make transaction
               const { data: sendTransactionData, error } = await fetchWithBQ({
                 command: 'sendTransaction',
-                service: Wallet,
+                service: WalletService,
                 args: [walletId, amount, fee, address, memos],
               });
 
@@ -530,7 +531,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     generateMnemonic: build.mutation<string[], undefined>({
       query: () => ({
         command: 'generateMnemonic',
-        service: Wallet,
+        service: WalletService,
       }),
       transformResponse: (response: any) => response?.mnemonic,
     }),
@@ -538,7 +539,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     getPublicKeys: build.query<number[], undefined>({
       query: () => ({
         command: 'getPublicKeys',
-        service: Wallet,
+        service: WalletService,
       }),
       transformResponse: (response: any) => response?.publicKeyFingerprints,
       providesTags: (keys) =>
@@ -555,7 +556,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ fingerprint }) => ({
         command: 'deleteKey',
-        service: Wallet,
+        service: WalletService,
         args: [fingerprint],
       }),
       invalidatesTags: (_result, _error, { fingerprint }) => [
@@ -580,7 +581,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ fingerprint }) => ({
         command: 'checkDeleteKey',
-        service: Wallet,
+        service: WalletService,
         args: [fingerprint],
       }),
     }),
@@ -588,7 +589,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     deleteAllKeys: build.mutation<any, undefined>({
       query: () => ({
         command: 'deleteAllKeys',
-        service: Wallet,
+        service: WalletService,
       }),
       invalidatesTags: [
         { type: 'Keys', id: 'LIST' },
@@ -607,7 +608,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ fingerprint, type, filePath, host }) => ({
         command: 'logIn',
-        service: Wallet,
+        service: WalletService,
         args: [fingerprint, type, filePath, host],
       }),
       invalidatesTags: ['LoggedInFingerprint'],
@@ -622,7 +623,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ fingerprint, host }) => ({
         command: 'logInAndSkipImport',
-        service: Wallet,
+        service: WalletService,
         args: [fingerprint, host],
       }),
       invalidatesTags: ['LoggedInFingerprint'],
@@ -638,7 +639,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ fingerprint, filePath, host }) => ({
         command: 'logInAndImportBackup',
-        service: Wallet,
+        service: WalletService,
         args: [fingerprint, filePath, host],
       }),
     }),
@@ -652,7 +653,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ filePath, options }) => ({
         command: 'getBackupInfo',
-        service: Wallet,
+        service: WalletService,
         args: [filePath, options],
       }),
     }),
@@ -666,7 +667,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ filePath, fingerprint }) => ({
         command: 'getBackupInfoByFingerprint',
-        service: Wallet,
+        service: WalletService,
         args: [filePath, fingerprint],
       }),
     }),
@@ -680,7 +681,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ filePath, words }) => ({
         command: 'getBackupInfoByWords',
-        service: Wallet,
+        service: WalletService,
         args: [filePath, words],
       }),
     }),
@@ -700,7 +701,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ fingerprint }) => ({
         command: 'getPrivateKey',
-        service: Wallet,
+        service: WalletService,
         args: [fingerprint],
       }),
       transformResponse: (response: any) => response?.privateKey,
@@ -718,7 +719,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ walletId, start, end, sortKey, reverse }) => ({
         command: 'getTransactions',
-        service: Wallet,
+        service: WalletService,
         args: [walletId, start, end, sortKey, reverse],
       }),
       transformResponse: (response: any) => response?.transactions,
@@ -733,17 +734,17 @@ export const walletApi = apiWithTag.injectEndpoints({
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onCoinAdded',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getTransactions,
         },
         {
           command: 'onCoinRemoved',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getTransactions,
         },
         {
           command: 'onPendingTransaction',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getTransactions,
         },
       ]),
@@ -757,7 +758,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ walletId }) => ({
         command: 'getTransactionsCount',
-        service: Wallet,
+        service: WalletService,
         args: [walletId],
       }),
       transformResponse: (response: any) => response?.count,
@@ -765,17 +766,17 @@ export const walletApi = apiWithTag.injectEndpoints({
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onCoinAdded',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getTransactionsCount,
         },
         {
           command: 'onCoinRemoved',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getTransactionsCount,
         },
         {
           command: 'onPendingTransaction',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getTransactionsCount,
         },
       ]),
@@ -789,7 +790,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ walletId }) => ({
         command: 'getNextAddress',
-        service: Wallet,
+        service: WalletService,
         args: [walletId, false],
       }),
       transformResponse: (response: any) => response?.address,
@@ -805,7 +806,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ walletId, newAddress }) => ({
         command: 'getNextAddress',
-        service: Wallet,
+        service: WalletService,
         args: [walletId, newAddress],
       }),
       transformResponse: (response: any) => response?.address,
@@ -820,7 +821,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ address }) => ({
         command: 'farmBlock',
-        service: Wallet,
+        service: WalletService,
         args: [address],
       }),
     }),
@@ -828,18 +829,18 @@ export const walletApi = apiWithTag.injectEndpoints({
     getHeightInfo: build.query<number, undefined>({
       query: () => ({
         command: 'getHeightInfo',
-        service: Wallet,
+        service: WalletService,
       }),
       transformResponse: (response: any) => response?.height,
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onSyncChanged',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getHeightInfo,
         },
         {
           command: 'onNewBlock',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getHeightInfo,
         },
       ]),
@@ -848,13 +849,13 @@ export const walletApi = apiWithTag.injectEndpoints({
     getCurrentDerivationIndex: build.query<number, undefined>({
       query: () => ({
         command: 'getCurrentDerivationIndex',
-        service: Wallet,
+        service: WalletService,
       }),
       providesTags: (result) => (result ? [{ type: 'DerivationIndex' }] : []),
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onNewDerivationIndex',
-          service: Wallet,
+          service: WalletService,
           onUpdate: (draft, data) => {
             draft.index = data?.additionalData?.index;
           },
@@ -869,7 +870,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ index }) => ({
         command: 'extendDerivationIndex',
-        service: Wallet,
+        service: WalletService,
         args: [index],
       }),
       invalidatesTags: [{ type: 'DerivationIndex' }],
@@ -878,24 +879,24 @@ export const walletApi = apiWithTag.injectEndpoints({
     getNetworkInfo: build.query<any, undefined>({
       query: () => ({
         command: 'getNetworkInfo',
-        service: Wallet,
+        service: WalletService,
       }),
     }),
 
     getSyncStatus: build.query<any, undefined>({
       query: () => ({
         command: 'getSyncStatus',
-        service: Wallet,
+        service: WalletService,
       }),
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onSyncChanged',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getSyncStatus,
         },
         {
           command: 'onNewBlock',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getSyncStatus,
         },
       ]),
@@ -904,7 +905,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     getWalletConnections: build.query<WalletConnections[], undefined>({
       query: () => ({
         command: 'getConnections',
-        service: Wallet,
+        service: WalletService,
       }),
       transformResponse: (response: any) => response?.connections,
       providesTags: (connections) =>
@@ -917,7 +918,7 @@ export const walletApi = apiWithTag.injectEndpoints({
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onConnections',
-          service: Wallet,
+          service: WalletService,
           onUpdate: (draft, data) => {
             // empty base array
             draft.splice(0);
@@ -937,7 +938,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ host, port }) => ({
         command: 'openConnection',
-        service: Wallet,
+        service: WalletService,
         args: [host, port],
       }),
       invalidatesTags: [{ type: 'WalletConnections', id: 'LIST' }],
@@ -950,7 +951,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ nodeId }) => ({
         command: 'closeConnection',
-        service: Wallet,
+        service: WalletService,
         args: [nodeId],
       }),
       invalidatesTags: (_result, _error, { nodeId }) => [
@@ -966,7 +967,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ filePath }) => ({
         command: 'createBackup',
-        service: Wallet,
+        service: WalletService,
         args: [filePath],
       }),
     }),
@@ -985,7 +986,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ start, end, sortKey, reverse, includeMyOffers, includeTakenOffers }) => ({
         command: 'getAllOffers',
-        service: Wallet,
+        service: WalletService,
         args: [start, end, sortKey, reverse, includeMyOffers, includeTakenOffers],
       }),
       transformResponse: (response: any) => {
@@ -1008,17 +1009,17 @@ export const walletApi = apiWithTag.injectEndpoints({
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onCoinAdded',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getAllOffers,
         },
         {
           command: 'onCoinRemoved',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getAllOffers,
         },
         {
           command: 'onPendingTransaction',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getAllOffers,
         },
       ]),
@@ -1027,7 +1028,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     getOffersCount: build.query<OfferCounts, undefined>({
       query: () => ({
         command: 'getOffersCount',
-        service: Wallet,
+        service: WalletService,
       }),
       providesTags: ['OfferCounts'],
     }),
@@ -1044,7 +1045,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ walletIdsAndAmounts, feeInMojos, driverDict, validateOnly, disableJSONFormatting }) => ({
         command: 'createOfferForIds',
-        service: Wallet,
+        service: WalletService,
         args: [walletIdsAndAmounts, feeInMojos, driverDict, validateOnly, disableJSONFormatting],
       }),
       invalidatesTags: [{ type: 'OfferTradeRecord', id: 'LIST' }, 'OfferCounts'],
@@ -1060,7 +1061,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ tradeId, secure, fee }) => ({
         command: 'cancelOffer',
-        service: Wallet,
+        service: WalletService,
         args: [tradeId, secure, fee],
       }),
       invalidatesTags: (result, error, { tradeId }) => [{ type: 'OfferTradeRecord', id: tradeId }],
@@ -1069,7 +1070,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     checkOfferValidity: build.mutation<any, string>({
       query: (offerData: string) => ({
         command: 'checkOfferValidity',
-        service: Wallet,
+        service: WalletService,
         args: [offerData],
       }),
     }),
@@ -1083,7 +1084,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ offer, fee }) => ({
         command: 'takeOffer',
-        service: Wallet,
+        service: WalletService,
         args: [offer, fee],
       }),
       invalidatesTags: [{ type: 'OfferTradeRecord', id: 'LIST' }, 'OfferCounts'],
@@ -1092,7 +1093,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     getOfferSummary: build.mutation<any, string>({
       query: (offerData: string) => ({
         command: 'getOfferSummary',
-        service: Wallet,
+        service: WalletService,
         args: [offerData],
       }),
     }),
@@ -1100,7 +1101,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     getOfferData: build.mutation<any, string>({
       query: (offerId: string) => ({
         command: 'getOfferData',
-        service: Wallet,
+        service: WalletService,
         args: [offerId],
       }),
     }),
@@ -1108,7 +1109,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     getOfferRecord: build.mutation<any, OfferTradeRecord>({
       query: (offerId: string) => ({
         command: 'getOfferRecord',
-        service: Wallet,
+        service: WalletService,
         args: [offerId],
       }),
     }),
@@ -1308,7 +1309,7 @@ export const walletApi = apiWithTag.injectEndpoints({
                 subscribeResponse = await baseQuery(
                   {
                     command: 'onTransactionUpdate',
-                    service: Wallet,
+                    service: WalletService,
                     args: [
                       (data: any) => {
                         const {
@@ -1512,7 +1513,7 @@ export const walletApi = apiWithTag.injectEndpoints({
             (async () => {
               const { data, error } = await fetchWithBQ({
                 command: 'getWallets',
-                service: Wallet,
+                service: WalletService,
               });
 
               if (error) {
@@ -1556,7 +1557,7 @@ export const walletApi = apiWithTag.injectEndpoints({
               poolWallets.map(async (wallet) => {
                 const { data, error } = await fetchWithBQ({
                   command: 'getPwStatus',
-                  service: Wallet,
+                  service: WalletService,
                   args: [wallet.id],
                 });
 
@@ -1574,7 +1575,7 @@ export const walletApi = apiWithTag.injectEndpoints({
               poolWallets.map(async (wallet) => {
                 const { data, error } = await fetchWithBQ({
                   command: 'getWalletBalance',
-                  service: Wallet,
+                  service: WalletService,
                   args: [wallet.id],
                 });
 
@@ -1730,7 +1731,7 @@ export const walletApi = apiWithTag.injectEndpoints({
         try {
           const { data, error } = await fetchWithBQ({
             command: 'getWallets',
-            service: Wallet,
+            service: WalletService,
           });
 
           if (error) {
@@ -1780,7 +1781,7 @@ export const walletApi = apiWithTag.injectEndpoints({
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onWalletCreated',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getWallets,
         },
       ]),
@@ -1960,7 +1961,7 @@ export const walletApi = apiWithTag.injectEndpoints({
       onCacheEntryAdded: onCacheEntryAddedInvalidate(baseQuery, [
         {
           command: 'onWalletCreated',
-          service: Wallet,
+          service: WalletService,
           endpoint: () => walletApi.endpoints.getNFTWalletsWithDIDs,
         },
       ]),
@@ -2087,7 +2088,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ address, message }) => ({
         command: 'signMessageByAddress',
-        service: Wallet,
+        service: WalletService,
         args: [address, message],
       }),
     }),
@@ -2101,7 +2102,7 @@ export const walletApi = apiWithTag.injectEndpoints({
     >({
       query: ({ id, message }) => ({
         command: 'signMessageById',
-        service: Wallet,
+        service: WalletService,
         args: [id, message],
       }),
     }),
