+++ packages/api/src/services/index.ts
@@ -3,6 +3,6 @@ export { default as Events } from './Events';
 export { default as Farmer } from './Farmer';
 export { default as FullNode } from './FullNode';
 export { default as Harvester } from './Harvester';
-export { default as Plotter } from './Plotter';
+export { default as PlotterService } from './PlotterService';
 export { default as Service } from './Service';
-export { default as Wallet } from './Wallet';
+export { default as WalletService } from './WalletService';
