+++ packages/api/src/wallets/Pool.ts
@@ -1,4 +1,4 @@
-import Wallet from '../services/Wallet';
+import Wallet from '../services/WalletService';
 
 export default class PoolWallet extends Wallet {
   async createNewWallet(initialTargetState: Object, fee: string) {
