+++ packages/gui/src/components/offers/OfferEditorConditionsPanel.tsx
@@ -1,4 +1,4 @@
-import { Wallet, WalletType } from '@chia-network/api';
+import type { Wallet, WalletType } from '@chia-network/api';
 import { useGetWalletBalanceQuery, useGetWalletsQuery } from '@chia-network/api-react';
 import {
   Amount,
